test: add test for list_videos
